significant clean-up & rework of async lock use for EMPRO state transitions
