+++ portal/tasks.py
@@ -15,7 +15,7 @@ from traceback import format_exc
 from celery.utils.log import get_task_logger
 from flask import current_app
 import redis
-from requests import Request, Session, post
+from requests import Request, Session
 from requests.exceptions import RequestException
 from sqlalchemy import and_
 
@@ -24,7 +24,6 @@ from .factories.app import create_app
 from .factories.celery import create_celery
 from .models.communication import Communication
 from .models.communication_request import queue_outstanding_messages
-from .models.observation import Observation
 from .models.qb_status import QB_Status
 from .models.qb_timeline import invalidate_users_QBT, update_users_QBT
 from .models.reporting import (
@@ -362,36 +361,8 @@ def celery_beat_health_check(**kwargs):
 @celery.task(name="tasks.extract_observations_task", queue=LOW_PRIORITY)
 def extract_observations_task(questionnaire_response_id):
     """Task wrapper for extract_observations"""
+    from portal.trigger_states.empro_states import extract_observations
     extract_observations(questionnaire_response_id)
-
-
-def extract_observations(questionnaire_response_id):
-    """Format and submit QuestionnaireResponse to SDC service; store returned Observations"""
-    from .models.questionnaire_response import QuestionnaireResponse
-    from .trigger_states.empro_states import (
-        enter_user_trigger_critical_section,
-        evaluate_triggers,
-        fire_trigger_events,
-    )
-    qnr = QuestionnaireResponse.query.get(questionnaire_response_id)
-
-    enter_user_trigger_critical_section(user_id=qnr.subject_id)
-
-    qnr_json = qnr.as_sdc_fhir()
-
-    SDC_BASE_URL = current_app.config['SDC_BASE_URL']
-    response = post(f"{SDC_BASE_URL}/$extract", json=qnr_json)
-    response.raise_for_status()
-
-    Observation.parse_obs_bundle(response.json())
-    db.session.commit()
-
-    # As scoring is complete, pass baton to evaluation process
-    # which also frees the locked critical section on completion
-    evaluate_triggers(qnr)
-
-    # Finally, fire any outstanding actions
-    fire_trigger_events()
 
 
 @celery.task(queue=LOW_PRIORITY)
